requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Google Services
google-cloud-firestore>=2.14.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.1.0

# Testing
pytest>=7.4.0
httpx>=0.26.0
